chore: split tests and add more profile tests
