+++ apps/web/src/components/Profile/Feed.tsx
@@ -117,7 +117,10 @@ const Feed: FC<FeedProps> = ({ profile, type }) => {
   }
 
   return (
-    <Card className="divide-y-[1px] dark:divide-gray-700">
+    <Card
+      className="divide-y-[1px] dark:divide-gray-700"
+      dataTestId={`profile-feed-type-${type.toLowerCase()}`}
+    >
       {publications?.map((publication, index) => (
         <SinglePublication
           key={`${publication.id}_${index}`}
