+++ apps/web/src/components/Profile/index.tsx
@@ -98,7 +98,6 @@ const ViewProfile: NextPage = () => {
           setShowFollowModal={setShowFollowModal}
         />
       </Modal>
-
       {profile?.name ? (
         <MetaTags title={`${profile?.name} (@${formatHandle(profile?.handle)}) • ${APP_NAME}`} />
       ) : (
