+++ packages/ui/src/TabButton.tsx
@@ -27,6 +27,7 @@ const TabButton: FC<TabButtonProps> = ({ name, icon, active, type, showOnSm = fa
         { 'text-brand bg-brand-100 bg-opacity-100 dark:bg-opacity-20': active },
         'hover:bg-brand-100 flex items-center space-x-2 rounded-lg px-4 py-2 text-sm font-medium text-gray-500 hover:bg-opacity-100 dark:hover:bg-opacity-20 sm:px-3 sm:py-1.5'
       )}
+      data-testid={`tab-button-${name.toLowerCase()}`}
       aria-label={name}
     >
       {icon}
